Separating GMod IO code from runtime behaviours.
