+++ ab3d2_source/c/version.h
@@ -2,5 +2,5 @@
 #   define TKG_VERSION 1
 #   define TKG_REVISION 1
 #   define TKG_BUILD 0
-#   define TKG_HASH "e3363c3"
+#   define TKG_HASH "817b791"
 #endif
